requirements.txt
# AI Image Detector 프로젝트 의존성 패키지

# 딥러닝 프레임워크
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# 데이터 처리
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
opencv-python>=4.8.0

# 시각화 및 분석
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# 설정 관리
pyyaml>=6.0

# 웹 프레임워크
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
python-multipart>=0.0.6
plotly>=5.15.0

# 유틸리티
tqdm>=4.65.0
tensorboard>=2.13.0

